refactor(commands): extract skill versioning into version_cmds.rs
skill_cmds.rs sat at 515 lines — still under the 600-line ceiling but
the third of the three files flagged by the architecture audit for
splitting. Unlike workspace_cmds.rs, skill_cmds has 8 distinct
subcommands with weaker internal coupling, so the split lines here are
more defensible on cohesion grounds than on raw size.

The three version-history subcommands form a tight cluster:

  cmd_release   (34 L)  git tag for `release/<skill>/<version>`
  cmd_rollback  (57 L)  git checkout-path → commit to restore skill
  cmd_diff      (14 L)  git diff between two refs for a skill

All three share the same five helpers (validate_skill_name, map_arg,
map_git, ensure_skill_exists, maybe_autosync_or_queue) and the same
mental model ("operate on a skill's git history"). They're orthogonal
to the registry/projection/capture cluster that stays in skill_cmds.

Move them to src/commands/version_cmds.rs behind the same `impl App`
block — same pattern as Batch 3c.

Result:
  skill_cmds.rs    515 → 407 L  (−21%, matches the target_cmds split ratio)
  version_cmds.rs  new → 121 L

Cleanup beyond the move:
- Dropped DiffArgs/ReleaseArgs/RollbackArgs from skill_cmds.rs imports
  — they're only used by the versioning subcommands.
- `cargo fmt` trimmed one trailing blank line the removal exposed.
- All other skill_cmds imports stayed (grep-verified: the remaining
  5 commands still use Uuid, chrono::Utc, remove_path_if_exists,
  V3BindingRule, V3ProjectionInstance, and the rest of the helpers set).

Lesson carried from Batch 3c: when stripping imports during a split,
grep each symbol against the surviving methods first rather than
guessing — saved a round of build-fix churn this time.

Tested: make fmt-check, make lint (clippy -D warnings), make test
(75/75 pass). Integration suite that exercises release/rollback/diff
via the CLI dispatcher passes unchanged, confirming behavior preserved.
Scope-risk: narrow (3 files).
Reversibility: clean — inline version_cmds back into skill_cmds.
Confidence: high.
