+++ src/commands/mod.rs
@@ -5,6 +5,7 @@ mod projections;
 mod skill_cmds;
 mod sync_cmds;
 mod target_cmds;
+mod version_cmds;
 mod workspace_cmds;
 
 use anyhow::Result;
