+++ src/commands/workspace_cmds.rs
@@ -1,3 +1,4 @@
+use std::fs;
 use std::path::Path;
 
 use chrono::Utc;
